move handleTextInput event handler into src/event-handler.ts
