+++ src/event-handler.ts
@@ -1,0 +1,6 @@
+import { ChangeEvent } from "react";
+import { StringStateSetter } from "../types";
+
+export function handleTextInput(e: ChangeEvent, setTextInput: StringStateSetter) {
+    setTextInput(e.target.value);
+}
